Error handling implemented

+++ WebApp1/WebApplication1/Controllers/Transaction/TransactionController.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
+using System.Diagnostics;
 
 namespace WebApplication1.Controllers.Transaction
 {
@@ -28,34 +29,50 @@ namespace WebApplication1.Controllers.Transaction
         [HttpGet]
         public async Task<JsonResult> GetAllTransactions()
         {
-            if (db.Transactions != null)
+            if (db!=null)
             {
-                var transactions = await GetNormilizedTransactions();
-                var result = Json(new { data = transactions});
-                return result;
+                if(db.Transactions != null)
+                {
+                    var transactions = await GetNormilizedTransactions();
+                    var result = Json(new { data = transactions });
+                    return result;
+
+                }
+                else
+                {
+                    throw new NullReferenceException("Transactions set does not exsist in current TransactionContext");
+                }
             }
             else
             {
-                throw new NullReferenceException("Transactions set does not exsist in current TransactionContext");
+                throw new NullReferenceException("Transactions database context does not exist");
             }
         }
 
         private async Task<List<Models.Transaction>> GetNormilizedTransactions()
         {
-            if (db.Transactions != null)
+            if(db!= null)
             {
-                var transactions = db.Transactions.ToList();
-                foreach (var transaction in transactions)
+                if (db.Transactions != null)
                 {
-                    transaction.Product = db.Products.FirstOrDefault(product => product.ID == transaction.ProductId);
-                    transaction.User = await _userManager.FindByIdAsync(transaction.UserId);
+                    var transactions = db.Transactions.ToList();
+                    foreach (var transaction in transactions)
+                    {
+                        transaction.Product = db.Products.FirstOrDefault(product => product.ID == transaction.ProductId);
+                        transaction.User = await _userManager.FindByIdAsync(transaction.UserId);
+                    }
+                    return transactions;
                 }
-                return transactions;
+                else
+                {
+                    throw new NullReferenceException("Transactions set does not exsist in current TransactionContext");
+                }
             }
             else
             {
-                throw new NullReferenceException("Transactions set does not exsist in current TransactionContext");
+                throw new NullReferenceException("Transactions database context does not exist");
             }
+
         }
 
         [HttpPost]
@@ -80,10 +97,33 @@ namespace WebApplication1.Controllers.Transaction
         }
         public IActionResult Delete(int Id)
         {
-            var transaction = db.Transactions.First(transaction => transaction.TransactionId == Id);
-            db.Transactions.Remove(transaction);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if(db != null)
+            {
+                if(db.Transactions != null)
+                {
+                    var transaction = db.Transactions.First(transaction => transaction.TransactionId == Id);
+                    try
+                    {
+                        db.Transactions.Remove(transaction);
+                        db.SaveChanges();
+                    }
+                    catch(Exception e)
+                    {
+                        Debug.WriteLine($"Error {e.Message}. Stack trace: {e.StackTrace}");
+                    }
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    throw new NullReferenceException("Transactions set does not exsist in current TransactionContext");
+                }
+
+            }
+            else
+            {
+                throw new NullReferenceException("Transactions database context does not exist");
+            }
+
         }
     }
 }
